+++ src/features/infravision/shaders/canvas.js
@@ -1,0 +1,1 @@
+export class InfravisionCanvasShader extends ColorAdjustmentsSamplerShader {}